% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genomidi.R
\name{composeSymphony}
\alias{composeSymphony}
\title{composeSymphony}
\usage{
composeSymphony(symphony)
}
\arguments{
\item{symphony}{A object of class Symphony}
}
\value{
None.
}
\description{
This function composes a non-composed object of class Symphony. In other words, it sets the appropiate GRanges objects in the Elements field of the tracks in the TrackList argument of a Symphony object.
}
\examples{
composeSymphony(symphony)
}
