% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genomidi.R
\name{printSymphony}
\alias{printSymphony}
\title{printSymphony}
\usage{
printSymphony(symphony)
}
\arguments{
\item{symphony}{An already composed object of class Symphony (See the composeSymphony function)}
}
\value{
A DataFrame.
}
\description{
This function prints a Symphony object in text (.asc) format.
}
\examples{
d=printSymphony(symphony)
}
