% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genomidi.R
\name{printTrack}
\alias{printTrack}
\title{printTrack}
\usage{
printTrack(track)
}
\arguments{
\item{track}{A Track object.}
}
\value{
A DataFrame.
}
\description{
This function prints an individual Track object in text (.asc) format.
}
\examples{
d=printTrack(track)
}
