% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genomidi.R
\docType{class}
\name{Symphony-class}
\alias{Symphony-class}
\alias{Symphony}
\title{The Symphony object}
\arguments{
\item{Label}{A label for the Symphony object.}

\item{Chr}{Chromosome where the region of interest is located.}

\item{Start}{Start coordinate.}

\item{End}{End coordinate.}

\item{Tempo}{Numeric value representing the tempo of the genomic symphony.}

\item{TrackList}{A list of Track objects containing all the instruments for the symphony.}
}
\description{
This is an object that stores all the data needed to produce a genomic symphony.
}

\examples{
S=Symphony(
	Label="centromere-at-chr-4",
	Chr="chr4",
	Start=33311305,
	End=47004872,
	Tempo=100000,
	TrackList=list(T.ATAC,T.GENE,T.LINE,T.SINE,T.DNA,T.LTR))
}
