% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genomidi.R
\docType{class}
\name{Track-class}
\alias{Track-class}
\alias{Track}
\title{The Track object
This is an object that stores the all data required for matching a given musical instrument with the data of a given type of genomic annotation.}
\arguments{
\item{DB}{A genomicRanges object with numeric metadata.}

\item{Aes}{A list of aethetics used to map the numeric metadata to a sound. Supported aesthetics are Name, Note and Vol.}

\item{Name}{A string with a name for the track.}

\item{Channel}{A numeric value indicating the MIDI channel to use.}

\item{Instrument}{Numeric value indicating the MIDI instrument to use.}

\item{Elements}{A genomicRanges object ready for producing sound from it (See the composeSymphony() function).}
}
\description{
The Track object
This is an object that stores the all data required for matching a given musical instrument with the data of a given type of genomic annotation.
}

\examples{
T.GENE=Track(
DB=GENEs.gr,
Aes=list(Name="names", Note="note", Vol="fixed_vol"),
Name="GENE",
Channel=5,
Instrument=53 # Voice Ooh
)
}
